PSQLADM-69 : Reading multiple cluster name from scheduler if write hostgroup matches.
Issue:
When we search based on hostgroup ID we were using a regular expression, but it
wasn't terminated properly, we were doing things like '20%' rather than '20 %'.
Thus we would match anything that started with '20', such as '200', '201', etc...

Solution:
Add a space after the hostgroup ID.  When we write out the write hostgroup ID, it
is part of a larger string and will have a space after it.