PSQLADM-97 : Tests don't work with PXC 5.6
Issue:
The tests fail in various places.  Ther are several reasons.
The first reason is that some of the tests disable the nodes
using pxc_maint_mode which is not supported in PXC 5.6. The
second reason is the differences in permissions/passwords
between 5.6 and 5.7.

Solution:
Add a version check and skip any test using pxc_maint_mode.
Also initialize the nodes differently (for accounts/permissions)
for 5.6