PSQLADM-94 : With PSQLADM-12 fix, node monitoring log is not showing cluster status
Issue:
The original code relied on a broken if condition.  I fixed the if condition
which caused a change in the underlying behavior.  The original code would
prit out the status of all nodes in the cluster, not just the ones that
were added/removed.

Solution:
Fixed the original if condition, and added code to print out the state
of nodes that were already in the cluster.
Also added code to ensure that debugging output is not lost.