PSQLADM-102 : v.1.4.10-dev branch: proxysql_galera_checker fails to upgrade old scheduler rows
Issue
Changes to the proxysql_galera_checker caused the loop over the
scheduler rows to fail.

Solution
Modified the loop to work the same way as the rest of the code.
Also parsed the options before the upgrade so that the upgrade
code can write it's output to the log.