PSQLADM-68 : Make scripts IPv6 compatible
Issue
Scripts are not compatible with IPv6 formatted addresses.
The major problem is that we parse the IP addr and port
by breaking on the ':', which is a problem for IPv6 addresses
since they include ':' in the address.

Solution
Fixup the places where we use addresses.
Also add an "--ipv6" option to the test scripts.