PSQLADM-120 : /bin/proxysql_galera_checker: line 514: $7: unbound variable
Issue
(Initial reason is not quite clear). But this is probably coming up becase
we did not enclose one of the parameters within quotes.  Which would cause
bash to not include it as a parameter if the value is missing.

Solution
Enclose the parameter within quotes always.
Thus
  "$port"
Instead of just
  $port
Does not fix the underlying problem, which seems to be the failure to
connect to proxysql, but should remove this error message from appearing.