PSQLADM-126 : Script is showing an error when --syncusers is invoked
Issue
When working with a proxysql instance wiht no entries in the mysql_users,
an error would be reported (since we weren't getting any values back
from proxysql).

Note that the syncusers would still function correctly.

Solution
Move the check of the return value to the query (the query will succeed
but return no data).  The error occurs in the pipeline processing of
the query result.