PSQLADM-129 : Stopping/Restarting ProxySQL can lead to multiple instances running at the same time
Issue
Stopping/restarting proxysql can cause multiple instances of proxysql_galera_checker to run
at the same time.  This occurs because one script runs normally (using cluster_name_galera_check.pid).
The other script fails to find the cluster name (because ProxySQL is down) and thus uses
the plain galera_check.pid.

Solution
If we fail to contact ProxySQL, we should just abort the script, as we would be continuing
with erroneous information (no matter where we are in the ProxySQL script).