PSQLADM-130 : proxysql_galera_checker script should consider user's custom max_connections setting
Issue
When inserting a row into the mysql_servers table, we would not set the
value for max_connections.  Thus the default value of 1000 would be used.
This causes problems with customers that need a higher default value.

Solution
Add an option to set the value of max_connections.  This can be
done on the command line and in the config file.

Also added another testsuite to test this settings (as much as
possible).

Also updated the version to 1.4.13.1