PSQLADM-219 : Scheduler is not respecting pxc_maint_transition_period
Issue
If a node is being shutdown, it sets pxc_maint_mode to 'SHUTDOWN',
however proxysql_galera_checker will set the state to OFFLINE_SOFT
on the first run.  On the second run, it will then move the node
to OFFLINE_HARD (ignoring the pxc_maint_transition_period).

Solution
Only move the node to OFFLINE_HARD if we are unable to connect
to the node, otherwise leave in OFFLINE_SOFT.