PSQLADM-327 : proxysql-admin --syncusers does not work for async replication
https://jira.percona.com/browse/PSQLADM-327

Issue
Running --syncusers or --sync-multi-cluster-users does not work if
the node being synced is not a cluster node.  The code checks to
see if the node is part of the cluster and errors out otherwise.

Solution
Added the --server option. If this is specified with --syncusers or
--sync-multi-cluster-users, then the check for cluster membership
is skipped, thus allowing standalone nodes to be synced.