PSQLADM-384: proxysql-admin: --write-node fails if cluster is read-only
https://jira.percona.com/browse/PSQLADM-384

Problem
-------
When running --write-node option along with --enable, prevented the
setting up proxysql due to read-only node checks even when --force was
specified. This affected operators.

Solution
--------
If --write-node is used with the --force flag, then we log a warning instead of an error
if the write-node is a read-only node.

In addition to the above problem, the following things have also been
changes.
1. Reordered options in help section
2. Corrected the logic for checking lockFilePath in
   percona-scheduler-admin.
3. Updated README.md
4. Added bats test
5. Added appropriate comments in config.toml