PSQLADM-511: proxysql-admin syncusers hangs at 1000 DB user or above
https://perconadev.atlassian.net/browse/PSQLADM-511

Problem
-------
Logic used by --syncusers is of O(n^2). When it tries to create a new
user in proxysql, it compares the new user to be added against all the
existing users thereby resulting in a huge delay.

Solution
--------
Modified the logic of --syncusers to use assosciative arrays to perform
comparison of MySQL and ProxySQL users.

The old implementation took about 6 mins to sync 500 users and 51 mins
to sync 1000 users.

The new implementation takes about 12 seconds to sync 500 users and 25
seconds to sync 1000 users.

In addition to the above,
- Moved the syncusers() function to proxysql-admin-common so that the same
logic will be used both by the proxysql-admin and
percona-scheduler-admin scripts.

- Introduces a new bats test to test the new improvement.